My-Blog-Serve/requirements.txt
aiodns==3.0.0
aiosmtpd==1.4.4.post2
altgraph==0.17.2
asgiref==3.6.0
atpublic==3.1.1
attrs==23.1.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.12
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cycler==0.11.0
Django==4.1.5
django-cors-headers==3.14.0
django-paginator==0.2a3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
fonttools==4.30.0
future==0.18.2
idna==3.3
itypes==1.2.0
jieba==0.42.1
Jinja2==3.1.2
kiwisolver==1.3.2
loguru==0.7.0
MarkupSafe==2.1.2
matplotlib==3.5.1
mysqlclient==2.1.1
numpy==1.22.3
openpyxl==3.1.2
packaging==21.3
pandas==1.4.2
pefile==2021.9.3
Pillow==9.0.1
pycares==4.3.0
pycparser==2.21
pyinstaller==4.7
pyinstaller-hooks-contrib==2021.4
PyJWT==2.6.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2022.1
pywin32-ctypes==0.2.0
qiniu==7.10.0
requests==2.27.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.8
verify-email==2.4.3
win32-setctime==1.1.0
you-get==0.4.1650
